+++ UnoApp6/Presentation/ErrorMessage.xaml
@@ -1,0 +1,48 @@
+﻿<Page   
+        xmlns:uen="using:Uno.Extensions.Navigation.UI" 
+        xmlns:uer="using:Uno.Extensions.Reactive.UI"
+        xmlns:utu="using:Uno.Toolkit.UI" 
+        xmlns:muxc="using:Microsoft.UI.Xaml.Controls" 
+        xmlns:um="using:Uno.Material" 
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"   
+        x:Class="UnoApp6.Presentation.ErrorMessage"  
+        mc:Ignorable="d"
+        Background="{ThemeResource BackgroundBrush}" 
+        NavigationCacheMode="Required"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
+
+    <Grid>
+        <TextBlock x:Name="Fehlermeldung" Text="ErrorMessage" />
+
+        <!--Darstellung Dateiname -->
+        <utu:AutoLayout Background="{ThemeResource BackgroundBrush}" Spacing="1">
+            <utu:AutoLayout Background="{ThemeResource PrimaryBrush}" Padding="153,12" PrimaryAxisAlignment="Center" Orientation="Horizontal">
+                <TextBlock Text="Fehler!" utu:AutoLayout.CounterAlignment="Center" Foreground="{ThemeResource OnSecondaryContainerBrush}"/>
+            </utu:AutoLayout>
+
+            <!--Darstellung der Seite -->
+            <utu:AutoLayout PrimaryAxisAlignment="Center" utu:AutoLayout.PrimaryAlignment="Stretch">
+
+                <!--Darstellung der JSON -->
+                <ScrollViewer utu:AutoLayout.PrimaryAlignment="Stretch" Height="673">
+                    <utu:AutoLayout Padding="600, 0" PrimaryAxisAlignment="Center">
+                        <TextBlock Text="" utu:AutoLayout.PrimaryAlignment="Stretch" Foreground="{ThemeResource OnSurfaceBrush}" Name="Error" />
+                    </utu:AutoLayout>
+                </ScrollViewer>
+
+                <utu:AutoLayout Background="{ThemeResource PrimaryBrush}" Padding="24,0,16,0" Justify="SpaceBetween" Orientation="Horizontal" Height="44">
+
+                    <!--Zurück Button -->
+                    <utu:AutoLayout utu:AutoLayout.CounterAlignment="Center" Width="24" Height="24">
+                        <HyperlinkButton Click="GoBack">
+                            <Path Data="F1 M 22 0 L 7 0 C 6.310000002384186 0 5.769999861717224 0.35000014305114746 5.409999847412109 0.880000114440918 L 0 9 L 5.409999847412109 17.110000610351562 C 5.769999861717224 17.640000581741333 6.310000002384186 18 7 18 L 22 18 C 23.100000023841858 18 24 17.100000023841858 24 16 L 24 2 C 24 0.8999999761581421 23.100000023841858 0 22 0 Z M 19 12.59000015258789 L 17.59000015258789 14 L 14 10.40999984741211 L 10.40999984741211 14 L 9 12.59000015258789 L 12.59000015258789 9 L 9 5.409999847412109 L 10.40999984741211 4 L 14 7.590000152587891 L 17.59000015258789 4 L 19 5.409999847412109 L 15.40999984741211 9 L 19 12.59000015258789 Z" Fill="{ThemeResource OnSurfaceBrush}" utu:AutoLayout.IsIndependentLayout="True" VerticalAlignment="Stretch" HorizontalAlignment="Stretch" Width="24" Height="18" />
+                        </HyperlinkButton>
+                    </utu:AutoLayout>
+
+                </utu:AutoLayout>
+            </utu:AutoLayout>
+        </utu:AutoLayout>
+    </Grid>
+</Page>
