+++ UnoApp6/Presentation/JSONListe2.xaml
@@ -9,7 +9,6 @@ xmlns:uer="using:Uno.Extensions.Reactive.UI"
   x:Class="UnoApp6.Presentation.JSONListe2"  
   mc:Ignorable="d" Background="{ThemeResource BackgroundBrush}" 
   NavigationCacheMode="Required"
-  x:Name="objektNummer"
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
 
     <Grid>
@@ -18,7 +17,7 @@ xmlns:uer="using:Uno.Extensions.Reactive.UI"
         <!--Darstellung Objektnummer -->
         <utu:AutoLayout Background="{ThemeResource BackgroundBrush}" Spacing="1">
             <utu:AutoLayout Background="{ThemeResource PrimaryBrush}" Padding="153,12" PrimaryAxisAlignment="Center" Orientation="Horizontal">
-                <TextBlock Text="Objektauswahl" utu:AutoLayout.CounterAlignment="Center" Foreground="{ThemeResource OnSecondaryContainerBrush}" Name="dateiName"/>
+                <TextBlock Text="JSON hier speichern" utu:AutoLayout.CounterAlignment="Center" Foreground="{ThemeResource OnSecondaryContainerBrush}" Name="xxx"/>
             </utu:AutoLayout>
 
             <!--Darstellung der Seite -->
@@ -27,10 +26,9 @@ xmlns:uer="using:Uno.Extensions.Reactive.UI"
                 <!--Darstellung der JSON -->
                 <ScrollViewer utu:AutoLayout.PrimaryAlignment="Stretch" Height="673">
                     <utu:AutoLayout Padding="600, 0" PrimaryAxisAlignment="Center">
-                        <TextBlock Text="" utu:AutoLayout.PrimaryAlignment="Stretch" Foreground="{ThemeResource OnSurfaceBrush}" Name="jsonData" />
+                        <TextBlock Text="" utu:AutoLayout.PrimaryAlignment="Stretch" Foreground="{ThemeResource OnSurfaceBrush}" Name="jsonDataNewFinal" />
                     </utu:AutoLayout>
                 </ScrollViewer>
-
 
                 <!--Objekt Öffnen Button -->
                 <utu:AutoLayout Background="{ThemeResource PrimaryBrush}" CornerRadius="20" utu:AutoLayout.IsIndependentLayout="True" Margin="536,590.5,0,0" VerticalAlignment="Top" HorizontalAlignment="Left">
