+++ UnoApp6/UnoApp6.csproj
@@ -63,31 +63,28 @@
 	<ItemGroup>
 		<None Remove="Presentation\Andern.xaml" />
 		<None Remove="Presentation\Bestaetigen.xaml" />
+		<None Remove="Presentation\ErrorMessage.xaml" />
 		<None Remove="Presentation\JSONListe2.xaml" />
-		<None Remove="Presentation\JSONListe3.xaml" />
 		<None Remove="Presentation\Offnen.xaml" />
 	</ItemGroup>
 
 	<ItemGroup>
 		<UpToDateCheckInput Remove="Presentation\Andern.xaml" />
 		<UpToDateCheckInput Remove="Presentation\Bestaetigen.xaml" />
+		<UpToDateCheckInput Remove="Presentation\ErrorMessage.xaml" />
 		<UpToDateCheckInput Remove="Presentation\JSONListe.xaml" />
 		<UpToDateCheckInput Remove="Presentation\JSONListe2.xaml" />
-		<UpToDateCheckInput Remove="Presentation\JSONListe3.xaml" />
 		<UpToDateCheckInput Remove="Presentation\Offnen.xaml" />
 	</ItemGroup>
 
 	<ItemGroup>
-	  <None Include="..\.editorconfig" Link=".editorconfig" />
+		<None Include="..\.editorconfig" Link=".editorconfig" />
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Page Update="Presentation\JSONListe3.xaml">
-	    <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
-	  </Page>
-	  <Page Update="Presentation\JSONListe2.xaml">
-	    <XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
-	  </Page>
+		<Page Update="Presentation\JSONListe2.xaml">
+			<XamlRuntime>$(DefaultXamlRuntime)</XamlRuntime>
+		</Page>
 	</ItemGroup>
